Uses method from same place
